fix: 三模块化编译修复(本地 :fabric/:neoforge/:common compileKotlin 全过)
- fabric/neoforge 跨项目 sourceSets 访问加 evaluationDependsOn(":mc26.1")
  (Gradle 9 惰性配置下访问未配置项目报 "Extension sourceSets does not exist")
- YaclConfigScreen:ButtonOption.description 改直接 OptionDescription(非 lambda)
- YaclLoginScreen 构造 back 改 Screen?(配置屏入口 McScreens.current() 可为 null)
- YaclFavPickerScreen:error 可变属性避免 smart cast 失败(局部快照)
- withColor(0xFFFF5C5C) 超 Int 范围被推断为 Long → .toInt()(3 处)

+++ mc26.1/src/main/kotlin/io/github/cyf112233/musicmc/ui/yacl/YaclMusicScreen.kt
@@ -330,7 +330,7 @@ class YaclMusicScreen : Screen(Component.literal("MusicMC")) {
             if (err != null) {
                 // 轻提示:聊天栏输出错误(与 ChatLyricSender 同一套 API,已验证可编译)
                 Minecraft.getInstance().gui.getChat().addClientSystemMessage(
-                    Component.literal(err).withColor(0xFFFF5C5C),
+                    Component.literal(err).withColor(0xFFFF5C5C.toInt()),
                 )
             }
         }
